+++ src/func/index.ts
@@ -1,0 +1,2 @@
+export * from "./row";
+export * from "./table";
