+++ test/test.js
@@ -1,16 +1,17 @@
-import {Dsm} from '../dist'
+import { Dsm } from "../dist";
+import { addRowAbove, addRowsAbove } from "../dist/func";
 
-console.clear()
+console.clear();
 
 const core = {
-tables:[],
-tablesSel:[],
-rowsSel:[]
-}
+  tables: [],
+  tablesSel: [],
+  rowsSel: [],
+};
 
-const ds = new Dsm({core})
+const ds = new Dsm({ core });
 
-console.log(ds.getModesReg('idle'))
+console.log(ds.getModesReg("idle"));
 
-	// ds.rows([], { select: 'test' });
-	// ds.rows([], { select: [0], which: 'all', place: 'place' });
+// ds.rows([], { select: 'test' });
+// ds.rows([], { select: [0], which: 'all', place: 'place' });
